Update Rtest.py
Is this a simple mistake in a loop that is intended to write each line to arcpy.AddMessage, not the entirety of the function's argument x to arcpy multiple times, for each separate line in x? Other examples in the code base show arcpy.AddMessage acting on single line strings.